requirements.txt
mysql-connector-python
mysqlclient